src/agile_methods.py
```python
class AgileMethods:
    def __init__(self):
        self.sprints = []
        self.backlog = []

    def create_sprint(self, start_date, end_date):
        sprint = {"start_date": start_date, "end_date": end_date, "tasks": []}
        self.sprints.append(sprint)

    def add_task_to_sprint(self, sprint_index, task):
        self.sprints[sprint_index]["tasks"].append(task)

    def move_task_to_backlog(self, sprint_index, task_index):
        task = self.sprints[sprint_index]["tasks"].pop(task_index)
        self.backlog.append(task)

    def move_task_to_sprint(self, backlog_index, sprint_index):
        task = self.backlog.pop(backlog_index)
        self.sprints[sprint_index]["tasks"].append(task)

    def get_sprint_report(self, sprint_index):
        return self.sprints[sprint_index]

    def get_backlog(self):
        return self.backlog
```